add HGCalTBCaloTrack object, tracking algorithm, HGCalTBCaloTrackProducer and python script + need to modify Reco/plugins/BuildFile.xml to use Eigen package

+++ Reco/python/hgcaltbclusterproducer_cfi.py
@@ -6,5 +6,5 @@ hgcaltbclusters = cms.EDProducer("HGCalTBClusterProducer",
                                  rechitCollection = cms.InputTag('hgcaltbrechits',"","unpack"),
                                  LayerZPositions= cms.untracked.vdouble(0.0, 4.67, 9.84, 14.27, 19.25, 20.4, 25.8, 31.4),#cern config 2
                                  #LayerZPositions= cms.untracked.vdouble(0.0, 5.35, 10.52, 14.44, 18.52, 19.67, 23.78, 25.92),#cern config 1
-                                  minEnergy = cms.untracked.double(100)
+                                 minEnergy = cms.untracked.double(9)
                                  )
